feat/MSSDK-2127: move options to usePrimer hook, use only one correct type from Primer library
